raspberry/requirements.txt
RPi.GPIO==0.7.1
requests==2.31.0
python-dotenv==1.0.0
pyngrok==7.0.0
psutil==5.9.5 